+++ kifi-backend/modules/shoebox/app/com/keepit/model/LibraryRepo.scala
@@ -38,7 +38,7 @@ trait LibraryRepo extends Repo[Library] with SeqNumberFunction[Library] {
   def pagePublished(page: Paginator)(implicit session: RSession): Seq[Library]
   def countPublished(implicit session: RSession): Int
   def filterPublishedByMemberCount(minCount: Int, limit: Int = 100)(implicit session: RSession): Seq[Library]
-  def pageByKind(kind: LibraryKind, page: Paginator)(implicit session: RSession): Seq[Library]
+  def pageByKindAndNoColor(kind: LibraryKind, page: Paginator)(implicit session: RSession): Seq[Library]
 }
 
 @Singleton
@@ -297,9 +297,9 @@ class LibraryRepoImpl @Inject() (
     } yield t).sortBy(_.updatedAt.desc).take(limit).list
   }
 
-  def pageByKind(kind: LibraryKind, page: Paginator)(implicit session: RSession): Seq[Library] = {
+  def pageByKindAndNoColor(kind: LibraryKind, page: Paginator)(implicit session: RSession): Seq[Library] = {
     (for {
-      t <- rows if t.kind === (LibraryKind.USER_CREATED: LibraryKind) && t.state === LibraryStates.ACTIVE
+      t <- rows if t.kind === (LibraryKind.USER_CREATED: LibraryKind) && t.state === LibraryStates.ACTIVE && t.color.isNull
     } yield t).drop(page.itemsToDrop).take(page.size).list
   }
 }
